% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.accelerometer}
\alias{plot.accelerometer}
\title{Plot accelerometer data}
\usage{
plot.accelerometer(accelerometer, pch = ".", type = "l")
}
\arguments{
\item{accelerometer}{Accelerometer data (e.g., data$AccelerometerDatum)}

\item{pch}{Plotting character.}

\item{type}{Line type.}
}
\description{
Plot accelerometer data
}

