% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.altitude}
\alias{plot.altitude}
\title{Plot altitude data.}
\usage{
plot.altitude(altitude, pch = ".", type = "l")
}
\arguments{
\item{altitude}{Altitude data (e.g., data$AltitudeDatum)}

\item{pch}{Plotting character.}

\item{type}{Line type.}
}
\description{
Plot altitude data.
}

