% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.battery}
\alias{plot.battery}
\title{Plot battery data.}
\usage{
plot.battery(battery, pch = ".", type = "l")
}
\arguments{
\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{batter}{Battery data (e.g., data$BatteryDatum)}
}
\description{
Plot battery data.
}

