% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.cell.tower}
\alias{plot.cell.tower}
\title{Plot cell tower data.}
\usage{
plot.cell.tower(cell.tower)
}
\arguments{
\item{cell.tower}{Cell tower data (e.g., data$CellTowerDatum)}
}
\description{
Plot cell tower data.
}

