% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.compass}
\alias{plot.compass}
\title{Plot compass data.}
\usage{
plot.compass(compass, pch = ".", type = "l")
}
\arguments{
\item{compass}{Compass data (e.g., data$CompassDatum)}

\item{pch}{Plotting character.}

\item{type}{Line type.}
}
\description{
Plot compass data.
}

