% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.light}
\alias{plot.light}
\title{Plot light data.}
\usage{
plot.light(light, pch = ".", type = "l")
}
\arguments{
\item{light}{Light data (e.g., data$LightDatum)}

\item{pch}{Plotting character.}

\item{type}{Line type.}
}
\description{
Plot light data.
}

