% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.location}
\alias{plot.location}
\title{Plot location data.}
\usage{
plot.location(location, resolution = "low")
}
\arguments{
\item{location}{Location data (e.g., data$LocationDatum)}

\item{resolution}{Mapping resolution ("low"/"high").}
}
\description{
Plot location data.
}

