% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.running.apps}
\alias{plot.running.apps}
\title{Plot running apps.}
\usage{
plot.running.apps(running.apps)
}
\arguments{
\item{running.apps}{Running apps data (e.g., data$RunningAppsDatum)}
}
\description{
Plot running apps.
}

