% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.screen}
\alias{plot.screen}
\title{Plot screen status.}
\usage{
plot.screen(screen)
}
\arguments{
\item{screen}{Screen status data (e.g., data$ScreenDatum)}
}
\description{
Plot screen status.
}

