% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.sound}
\alias{plot.sound}
\title{Plot sound data.}
\usage{
plot.sound(sound, pch = ".", type = "l")
}
\arguments{
\item{sound}{Sound data (e.g., data$SoundDatum)}

\item{pch}{Plotting character.}

\item{type}{Line type.}
}
\description{
Plot sound data.
}

