% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.SoundDatum}
\alias{plot.SoundDatum}
\title{Plot sound data.}
\usage{
\method{plot}{SoundDatum}(x, pch = ".", type = "l", ...)
}
\arguments{
\item{x}{Sound data.}

\item{pch}{Plotting character.}

\item{type}{Line type.}

\item{...}{Other plotting parameters.}
}
\description{
Plot sound data.
}
\examples{
data = read.sensus.json(system.file("extdata", "example.data.txt", package="SensusR"))
plot(data$SoundDatum)
}

