% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.speed}
\alias{plot.speed}
\title{Plot speed data.}
\usage{
plot.speed(speed, pch = ".", type = "l")
}
\arguments{
\item{speed}{Speed data (e.g., data$SpeedDatum)}

\item{pch}{Plotting character.}

\item{type}{Line type.}
}
\description{
Plot speed data.
}

