% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.telephony}
\alias{plot.telephony}
\title{Plot telephony data.}
\usage{
plot.telephony(telephony)
}
\arguments{
\item{telephony}{Telephony data (e.g., data$TelephonyDatum)}
}
\description{
Plot telephony data.
}

