% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{plot.timestamp.lags}
\alias{plot.timestamp.lags}
\title{Plot timestamp lags for a Sensus data frame.}
\usage{
plot.timestamp.lags(data, mfrow = c(4, 4))
}
\arguments{
\item{data}{Data to plot lags for (e.g., the result of \code{read.sensus.json}.)}
}
\description{
Plot timestamp lags for a Sensus data frame.
}

