% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\name{read.sensus.json}
\alias{read.sensus.json}
\title{Read JSON-formatted Sensus data.}
\usage{
read.sensus.json(path, convert.to.local.timezone = FALSE)
}
\arguments{
\item{path}{Path to JSON file.}

\item{convert.to.local.timezone}{Whether or not to convert timestamps to the local timezone.}
}
\value{
All data, listed by type.
}
\description{
Read JSON-formatted Sensus data.
}

