\name{SensusR-package}
\alias{SensusR-package}
\alias{SensusR}
\docType{package}
\title{
Sensus Analytics
SensusR
}
\description{
This package provides access and analytic functions for Sensus data.
}
\details{
\tabular{ll}{
Package: \tab SensusR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-17\cr
License: \tab GPL-3\cr
}
This package provides access and analytic functions for Sensus data.
}
\author{
Matthew S. Gerber

Maintainer: Matthew S. Gerber <gerber.matthew@gmail.com>
}
\references{
}
\keyword{ package }
\keyword{ Sensus }
\keyword{ Sensing }
\seealso{
}
\examples{
}
