% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SensusR.R
\docType{package}
\name{SensusR}
\alias{SensusR}
\alias{SensusR-package}
\title{SensusR:  Sensus Analytics}
\description{
This package provides access and analytic functions for Sensus data.
}
\section{SensusR functions}{

The SensusR functions handle reading, cleaning, plotting, and otherwise analyzing data collected
via the Sensus system.
}

